+++ object_detection/object_detection_v1/.gitignore
@@ -1,0 +1,2 @@
+# Ignores from object_detection_v1
+/scripts_old
