--- object_detection/size_estimator/.gitignore
@@ -1,2 +1,0 @@
-# Ignores from size estimator
-fromTobben.py
