+++ object_detection/size_estimator_deprecated/package.xml
